Move ReceiverFactory to "receiver" package.
This is part 1 of 3 of the refactoring plan to eliminate factories.go
and move the content to corresponding packages (receiver, exporter, processor).
