+++ receiver/opencensusreceiver/factory.go
@@ -19,14 +19,13 @@ import (
 
 	"go.uber.org/zap"
 
-	"github.com/open-telemetry/opentelemetry-service/factories"
 	"github.com/open-telemetry/opentelemetry-service/models"
 
 	"github.com/open-telemetry/opentelemetry-service/consumer"
 	"github.com/open-telemetry/opentelemetry-service/receiver"
 )
 
-var _ = factories.RegisterReceiverFactory(&receiverFactory{})
+var _ = receiver.RegisterReceiverFactory(&receiverFactory{})
 
 const (
 	// The value of "type" key in configuration.
@@ -43,7 +42,7 @@ func (f *receiverFactory) Type() string {
 }
 
 // CustomUnmarshaler returns nil because we don't need custom unmarshaling for this config.
-func (f *receiverFactory) CustomUnmarshaler() factories.CustomUnmarshaler {
+func (f *receiverFactory) CustomUnmarshaler() receiver.CustomUnmarshaler {
 	return nil
 }
